crediscan/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
crawl4ai
aiohttp
tenacity